fix: implement review comments
